Time system
I've added a custom time system to the game, with 12 hour days, 5 day weeks and a 12 month calendar. There is a TIME command to display the current time. There are also scheduled scripts to display messages at a particular time. These will need to be plugged into the code somewhere to ensure that they start up, or they can be started in game with the @script command.